fix: hn28f101: Generate and map 1MHz clock to EXTAL
EXTAL and XTAL are two sides of a clock circuit. Per the H8/538F
hardware manual, you can either put an oscillator circuit between the
two of them, or drive EXTAL manually with an input PWM.

Here, do the latter; this uses a timer to provide a 1MHz output from
Pin 12 of the Mega2560.